fix: improve credential handling and UX
Handled password visibility toggle so that it no longer tries to hide a password if none is present. Sorted credential items in the welcome page in ascending alphabetical order. Trimmed all user inputs to remove leading/trailing spaces. Refreshed 'createdDate' of credential items when corrisponding passwords are updated. Showed message prompting user to add a new fingerprint if not already done so.